requirements.txt
langchain==0.3.12
google-generativeai==0.8.3
langchain-google-genai==2.0.7 
PyPDF2==3.0.1
python-dotenv==1.0.1
pypdf==5.1.0
langchain-community==0.3.12
sentence-transformers==3.3.1
langchainhub==0.1.21
streamlit==1.41.1
cohere==5.13.3
ipywidgets==8.1.5 
faiss-cpu==1.9.0.post1
